fix bug xoa khi khong chon doi tuong

+++ DoAn-2/MenuTab/frmKhachHang.cs
@@ -144,13 +144,14 @@ namespace restaurant_management.MenuTab
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("Bạn chắc chắn muốn xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
-            {
-                if (KhachHangDAO.Instance.DeleteKhachHang(int.Parse(txtMaKhachHang.Text)))
-                    MessageBox.Show("Xoá thành công!");
-                SetBtnEdit_Off();
-                LoadData();
-            }
+            if(!string.IsNullOrEmpty(txtMaKhachHang.Text))
+                if (MessageBox.Show("Bạn chắc chắn muốn xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    if (KhachHangDAO.Instance.DeleteKhachHang(int.Parse(txtMaKhachHang.Text)))
+                        MessageBox.Show("Xoá thành công!");
+                    SetBtnEdit_Off();
+                    LoadData();
+                }
         }
 
         private void txtTimKiem_TextChanged(object sender, EventArgs e)
